Initial Drop - Network Issues
Backpropagation does not seem to be working. However, the overall logic
of this neural network is going to be replaced by simply pointing the
robot in the direction of the GPS coordinates and having A* algorithm
avoid objects. This neural network adds too much overhead for so few
overall options.